+++ ensure-no-alloc/src/main.rs
@@ -9,6 +9,7 @@ fn panic(_info: &core::panic::PanicInfo) -> ! {
 
 #[no_mangle]
 pub extern "C" fn _start() -> ! {
+    compat_test::check();
     loop {}
 }