+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Add no-alloc support.
+- Add no-alloc support, including parsing and formatting.
+- Replace the `String` error from `ByteSize::from_str()` with `ByteSizeParseError`.
 
 ## 2.7.0